pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
]

[project.scripts]
gmail-cli = "backend.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
